chore: refactor tasks to use a base task
This refactors the tasks to use a base task which provides a common ID
for the task, reducing the amount of boilerplate code needed for each
task.
